fix: improve filtering of elements in HideAdElementNano function (#107)
* fix: improve filtering of elements in HideAdElementNano function

This commit improves the filtering of elements in the HideAdElementNano function. It adds an additional filter to remove elements that have a specific class and contain an SVG path. This filter is applied to elements that have a margin-top greater than 10 pixels. This change is made to address an issue with the NamuNews Mobile layout.

* chore: update version

* build userscript